Refactor internalization into a package.
For readability.

No significant change on benchmark numbers.
