+++ src/nti/externalization/internalization/_externals.pxd
@@ -1,0 +1,12 @@
+# definitions for internalization.py
+import cython
+
+
+# imports
+cdef MutableSequenc
+cdef component
+cdef IExternalReferenceResolver
+
+# XXX: This is only public for testing
+cpdef resolve_externals(object_io, updating_object, externalObject,
+                        registry=*, context=*)
